requirements.txt
altair==5.0.1
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.5
cryptography==41.0.2
decorator==5.1.1
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
isodate==0.6.1
Jinja2==3.1.2
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.25.1
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
platformdirs==3.8.1
protobuf==4.23.4
pyarrow==12.0.1
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.7.0
Pympler==1.0.1
pyodbc==4.0.39
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
referencing==0.29.1
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rich==13.4.2
rpds-py==0.8.10
simple-salesforce==1.12.4
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.19
streamlit==1.24.1
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
zeep==4.2.1
zipp==3.16.1
